preserve bindist through USE filtering ======================================
Normally, anything not in an ebuild's IUSE is stripped from the USE
passed to ebuild.sh. This patch allows "bindist" to not be stripped,
so it can be used as a means to disable pre-merge sanity checks that
may exist in pkg_setup() and pkg_pretend() but will not otherwise
affect the resultant build.

If "bindist" will affect some functionality in the package, then it
should be added to an ebuild's IUSE.