Reverting commit 637a13355cf4cc2906e34b61606d31d62acc9fcc -
 USE_EXPAND: incremental expand for profiles

The problem with these changes is that they are riddled with bugs.
For example:

INPUT_DEVICES="-evdev" really doesn't remove evdev from INPUT_DEVICES
or input_devices_evdev from USE.
INPUT_DEVICES="-* evdev" sets the input_devices_-* USE variable :(

There is no real way to remove values once set, which was the
purpose of this feature in the first place. So Funtoo is reverting
back to the original behavior, which is that INPUT_DEVICES is a
non-incremental variable, which means any new definitions of
INPUT_DEVICES overrides the old definition, and things like -foo
and -* are treated as invalid and throw a warning.

Thanks to Jonas Bernoulli for pointing out this buggy behavior.